Added error config files + add treatment_common

+++ unit_test_m/tester.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
+EXE_NAME="loader.exe"
 CONF_PATH="./confs/*"
 INC_PATH="../inc"
 SRCS_PATH="../srcs/"
@@ -16,22 +17,27 @@ White='\033[0;37m'
 
 echo -e "To redirect the output of each test add 1 argument\n"
 echo -e "${Red}Compiling....${Reset}\n"
-clang++ -Wall -Wextra -I$INC_PATH -g -fsanitize=address ${SRCS_PATH}Loader.cpp ${SRCS_PATH}Server.cpp ${SRCS_PATH}Route.cpp ${SRCS_PATH}LoaderTreatment.cpp ./main_test_loader.cpp
+clang++ -Wall -Wextra -I$INC_PATH -g -fsanitize=address ${SRCS_PATH}Loader.cpp ${SRCS_PATH}Server.cpp ${SRCS_PATH}Route.cpp ${SRCS_PATH}LoaderTreatment.cpp ./main_test_loader.cpp -o ${EXE_NAME}
 
 i=1
-for f in $CONF_PATH
-do
-	echo -e "${Blue}Executing Test nº$i${Reset} - ${Red}FILE : $f${Reset}"
-	if [ $# -eq 1 ] 
-	then
-		./a.out $f > test_output/test_${i}.out
-	else
-		./a.out $f 
-	fi
-	((i=i+1))
-done
-if [ $# -eq 1 ]
+if test -f "./${EXE_NAME}"
 then
-	echo -e "\n${Green}The output of each tests has been redirected in a file test_<test number>.out in the test_output folder${Reset}"
+	for f in $CONF_PATH
+	do
+		echo -e "${Blue}Executing Test nº$i${Reset} - ${Red}FILE : $f${Reset}"
+		if [ $# -eq 1 ] 
+		then
+			./${EXE_NAME} $f > test_output/test_${i}.out
+		else
+			./${EXE_NAME} $f 
+		fi
+		((i=i+1))
+	done
+	if [ $# -eq 1 ]
+	then
+		echo -e "\n${Green}The output of each tests has been redirected in a file test_<test number>.out in the test_output folder$	{Reset}"
+	fi
+else
+	echo -e "\n${Red}Error: Compilation Failed${Reset}"
 fi
-rm -rf a.out a.out.dSYM
+rm -rf ${EXE_NAME} ${EXE_NAME}.dSYM
